pipeline/requirements.txt
# Requirements for My Juridic Assistant Pipeline
# Python 3.8+

# Core dependencies
python-dotenv==1.0.0

# For Supabase integration (Task #5)
supabase==2.3.4
postgrest==0.16.0

# For embeddings generation (Task #5)
openai==1.12.0

# For data processing
numpy==1.26.3

# Optional: for Voyage AI embeddings
# voyageai==0.2.0

# Optional: for Anthropic embeddings
# anthropic==0.18.1
